requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
defusedxml==0.7.1
dill==0.3.8
evaluate==0.4.3
filelock==3.13.1
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2024.6.1
fvcore==0.1.5.post20221221
huggingface-hub==0.33.0
idna==3.10
iopath==0.1.10
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.8
lightning-utilities==0.14.3
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.16
networkx==3.3
numpy==2.1.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.0.0
portalocker==3.2.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycocotools==2.0.10
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
simsimd==6.4.9
six==1.17.0
stringzilla==3.12.5
supervision==0.25.1
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.4
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchmetrics==1.7.3
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.156
ultralytics-thop==2.0.14
urllib3==2.5.0
xxhash==3.5.0
yacs==0.1.8
yarl==1.20.1
